Major Update
This is a massive update to the repository adding in ground breaking changes to some RT packets for Jak X: Combat Racing.
New packets:
`GetClanTeamChallenges` - Jak X
`RevokeClanTeamChallenges` - Jak X
`RespondToClanTeamChallenges` - Jak X
`RequestClanTeamChallenge` - Jak X
`UpdateClanLadderStatsWide_DeltaRequest/Response` - Jak X
`FileUpdateAuxMetaDataRequest` - Lemmings
`MatchCreateGameResponse` - Starhawk
`PartyJoinByIndexRequest/Response` - Starhawk
`DME Tokens` - PS Home
`Session Master 108` - Jak X

Added MAG packets

`NetAddress` Updated to support Binary iterations

+++ Server.Database/Models/ChannelDTO.cs
@@ -6,10 +6,10 @@
         public int AppId { get; set; }
         public string Name { get; set; }
         public int MaxPlayers { get; set; }
-        public int GenericField1 { get; set; }
-        public int GenericField2 { get; set; }
-        public int GenericField3 { get; set; }
-        public int GenericField4 { get; set; }
+        public ulong GenericField1 { get; set; }
+        public ulong GenericField2 { get; set; }
+        public ulong GenericField3 { get; set; }
+        public ulong GenericField4 { get; set; }
         public int GenericFieldFilter { get; set; }
     }
 }
